+++ R/utils.r
@@ -38,7 +38,7 @@ env_name <- function(x) {
 }
 
 find_first_srcref <- function(calls) {
-  for (call in rev(calls)) {
+  for (call in calls) {
     srcref <- attr(call, 'srcref')
     if (!is.null(srcref))
       return(srcref)