+++ NEWS.md
@@ -36,7 +36,7 @@
 
 * `make_expectation()` uses `expect_equal()`.
 
-* Errors now also get a srcref so you can find them more easily.
+* Errors and skips now also get a srcref so you can find them more easily.
 
 * `setup_test_dir()` has been removed. If you used it previously, instead use
   `source_test_helpers()` and `find_test_scripts()`.